fix: Move strategies outside of internal
