+++ providers/multi-provider/pkg/strategies/first_success.go
